Add new atoms for user state management
Introduced new atoms to manage user state in different scenarios. This includes maintaining the state of authentication, onboarded wallets, and provider clients (i.e., Arweave, Warp and Ethereum). These atoms are helpful in tracking signed and unsigned accounts and ensuring only accounts connected to supported chains are taken into account. They are also used to handle instances of Usher's Partnerships and Authentication and manage the utilities of user state,checking if the user is connected to unsigned or unsupported chains.